CmdFIFO was throwing away rpc requests too soon
Python rounding was causing CmdFIFO to throw away RPC requests early. Issue was conversion to an integer. Making sure both the min_ms_interval and timestamp are floats, fixes this.

+++ src/main/python/Host/Common/CmdFIFO.py
@@ -849,8 +849,8 @@ class CmdFIFOServerProxy(object):
 
         # Check to see if it is too soon to process this rpc.
         now = time.time()
-        dt_ms = int( (now - self.time_since_last_rpc)*1000 )
-        if dt_ms < self.min_ms_interval:
+        dt_ms = ((now - self.time_since_last_rpc) * 1000)
+        if dt_ms < float(self.min_ms_interval):
             return
         self.time_since_last_rpc = now
